osrd_schemas: rename and drop schema-sync check
osrd_schemas is now generated automatically (renamed from
osrd_schemas_auto), so there is no more hand-written schema to
keep in sync. remove the CI reminder that asked people to sync
and bump versions between osrd_schemas and editoast/schemas.
