+++ .github/dependabot.yml
@@ -209,13 +209,13 @@ updates:
     labels:
       - "dependencies"
   - package-ecosystem: "uv"
-    directory: "/osrd_schemas_auto/"
+    directory: "/osrd_schemas/"
     schedule:
       interval: "daily"
     cooldown:
       default-days: 5
     commit-message:
-      prefix: "osrd_schemas_auto:"
+      prefix: "osrd_schemas:"
     open-pull-requests-limit: 100
     labels:
       - "dependencies"
