+++ osrd_schemas/generate-types.sh
