+++ osrd_schemas/osrd_schemas/external_generated_inputs.py
