+++ osrd_schemas/osrd_schemas/infra_editor.py
