+++ osrd_schemas/osrd_schemas/models.py
