+++ osrd_schemas/osrd_schemas/switch_type.py
