+++ python/railjson_generator/railjson_generator/schema/infra/speed_section.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 from typing import List, Optional
 
-from osrd_schemas_auto import models
+from osrd_schemas import models
 
 from railjson_generator.schema.infra.range_elements import (
     ApplicableDirectionsTrackRange,
