+++ scripts/sync-infra-schema.sh
@@ -2,14 +2,14 @@
 
 set -euo pipefail
 
-OSRD_SCHEMA_DIR="$(realpath "$(dirname "$0")/../osrd_schemas_auto")"
+OSRD_SCHEMA_DIR="$(realpath "$(dirname "$0")/../osrd_schemas")"
 FRONT_DIR="$(realpath "$(dirname "$0")/../front")"
 
-echo "Syncing osrd_schemas_auto python dependencies"
+echo "Syncing osrd_schemas python dependencies"
 uv --directory "$OSRD_SCHEMA_DIR" sync
 
 echo "Generating the infra json schema"
-uv --directory "$OSRD_SCHEMA_DIR" run python -m osrd_schemas_auto.infra_editor >"$FRONT_DIR/src/reducers/osrdconf/infra_schema.json"
+uv --directory "$OSRD_SCHEMA_DIR" run python -m osrd_schemas.infra_editor >"$FRONT_DIR/src/reducers/osrdconf/infra_schema.json"
 
 echo "Extracting the infra editor English translations"
-uv --directory "$OSRD_SCHEMA_DIR" run python -m osrd_schemas_auto.infra_editor --translation >"$FRONT_DIR/public/locales/en/infraEditor.json"
+uv --directory "$OSRD_SCHEMA_DIR" run python -m osrd_schemas.infra_editor --translation >"$FRONT_DIR/public/locales/en/infraEditor.json"
